move `IndexNode` to `gelement-nodes`
