+++ node-graph/gelement-nodes/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod animation;
 pub mod blending_nodes;
 pub mod conversion;
+pub mod index;
 pub mod instance;
 pub mod logic;
 pub mod transform_nodes;
